Major refactor, easier maintainability
