+++ src/handlers/session.rs
@@ -16,11 +16,10 @@ use serenity::all::{
 };
 use tracing::{info, warn};
 
+use crate::models::server::*;
+use crate::models::session::*;
 use crate::database::Database;
 use crate::models::command::{CommandContext};
-use crate::models::player::Role;
-use crate::models::data::{ Group, Session, SessionPlayer, SessionStatus };
-use crate::models::data::Team;
 
 /// Checks if a user has the specified role.
 ///
@@ -132,13 +131,13 @@ async fn move_players_to_team_channels(
 //
 
 /// `/join` and `/leave`
-pub async fn queue<'a>(cc: &'a CommandContext<'a>) -> Result<()> {
+pub async fn queue<'a>(cc: &'a CommandContext<'a>, mut guild: Server) -> Result<()> {
     info!("Processing queue command from user {}", cc.intax.user.id);
     let client_id  = cc.intax.user.id;
     let channel_id = cc.intax.channel_id;
     
     // Get group of current channel
-    let mut group = cc.db.get_group_by_channel(channel_id).await?;
+    let group = guild.get_group(channel_id).unwrap();
     
     // Check if we have idle sessions
     match group.get_sessions_by_status(&SessionStatus::Idle).len() {
@@ -178,10 +177,10 @@ pub async fn queue<'a>(cc: &'a CommandContext<'a>) -> Result<()> {
 }
 
 /// `/status`
-pub async fn status<'a>(cc: &'a CommandContext<'a>) -> Result<()> {
+pub async fn status<'a>(cc: &'a CommandContext<'a>, mut guild: Server) -> Result<()> {
     info!("Processing queue status command");
     // Get active group with session
-    let group = cc.db.get_group_by_channel(cc.intax.channel_id).await?;
+    let group = guild.get_group(cc.intax.channel_id).unwrap();
 
     // If group has no sessions or session pool, return empty count
     let count = if group.sessions.is_empty() {
@@ -219,7 +218,7 @@ pub async fn status<'a>(cc: &'a CommandContext<'a>) -> Result<()> {
 }
 
 /// `/shuffle`
-pub async fn shuffle(cc: &CommandContext<'_>) -> Result<()> {
+pub async fn shuffle(cc: &CommandContext<'_>, mut guild: Server) -> Result<()> {
     info!("Processing shuffle command");
     // Check permissions
     if !check_role(cc, &Role::Runner).await? {
@@ -229,7 +228,7 @@ pub async fn shuffle(cc: &CommandContext<'_>) -> Result<()> {
 
     // Get active group with session
     // TODO: Replace 0 with the actual queue channel ID for the group you want
-    let group = cc.db.get_group_by_channel(ChannelId::new(0)).await?; // <-- FIX: supply correct queue_id
+    let group = guild.get_group(cc.intax.channel_id).unwrap();
 
     if group.sessions.is_empty() {
         cc.create_bot_reply("No active sessions.").await?;
@@ -291,11 +290,7 @@ pub async fn shuffle(cc: &CommandContext<'_>) -> Result<()> {
 }
 
 /// `/accept`
-pub async fn accept(cc: &CommandContext<'_>, session_id: &Option<String>) -> Result<()> {
-    info!(
-        "Processing accept command for session ID: {}",
-        session_id.clone().unwrap_or("None".to_string())
-    );
+pub async fn accept(cc: &CommandContext<'_>, mut guild: Server) -> Result<()> {
     // Check permissions
     if !check_role(cc, &Role::Runner).await? {
         cc.create_bot_reply("Only runners can accept sessions!").await?;
@@ -304,7 +299,7 @@ pub async fn accept(cc: &CommandContext<'_>, session_id: &Option<String>) -> Res
 
     // Get the group for the current channel
     let channel_id = cc.intax.channel_id;
-    let mut group = cc.db.get_group_by_channel(channel_id).await?;
+    let group = guild.get_group(channel_id).unwrap();
 
     match group.get_sessions_by_status(&SessionStatus::Hot).len() {
         0 => {
@@ -332,7 +327,7 @@ pub async fn accept(cc: &CommandContext<'_>, session_id: &Option<String>) -> Res
 /// `/end`
 ///
 /// * `cc` - The command context.
-pub async fn end(cc: &CommandContext<'_>) -> Result<()> {
+pub async fn end(cc: &CommandContext<'_>, mut guild: Server) -> Result<()> {
     info!("Processing end command");
     // Check permissions
     if !check_role(cc, &Role::Runner).await? {
@@ -340,7 +335,7 @@ pub async fn end(cc: &CommandContext<'_>) -> Result<()> {
         return Ok(());
     }
 
-    let mut group = cc.db.get_group_by_channel(cc.intax.channel_id).await?;
+    let group = guild.get_group(cc.intax.channel_id).unwrap();
 
     // Find the session that the current user is participating in
     let user_id = cc.intax.user.id;
