+++ src/models/mod.rs
@@ -1,12 +1,13 @@
 pub mod command;
 pub mod data;
-pub mod file;
+pub mod pug;
 pub mod manager;
 pub mod player;
 pub mod server;
 pub mod setup_state;
+pub mod session;
 
 pub use command::*;
 pub use data::*;
-pub use file::*;
+pub use pug::*;
 pub use player::*;
